feat(picture): add user permission check before deleting

+++ controller/picture.go
@@ -21,6 +21,29 @@ func DeletePicture(c echo.Context) error {
 			Message:    err.Error(),
 		})
 	}
+
+	var picture model.Picture
+	var memory model.Memory
+	currentUser, _ := utils.GetAuthUser(c)
+
+	if err := db.First(&picture, params.ID).Error; err != nil {
+		return utils.SendResponse(c, utils.BaseResponse{
+			StatusCode: http.StatusNotFound,
+			Message:    err.Error(),
+		})
+	}
+
+	// ignoring error because if the previous query succeeds
+	// it means that it's parent (the memory) exist and the db is working fine.
+	db.First(&memory, picture.MemoryID)
+
+	if memory.UserID != currentUser.UserID {
+		return utils.SendResponse(c, utils.BaseResponse{
+			StatusCode: http.StatusForbidden,
+			Message:    "You do not own this memory",
+		})
+	}
+
 	if err := db.Delete(&model.Picture{}, params.ID).Error; err != nil {
 		return utils.SendResponse(c, utils.BaseResponse{
 			StatusCode: http.StatusInternalServerError,
